Add files via upload
Pruefniveaus calulates by given size of a delivery(for example number of gears), exam level and
acceptable quality level the sample size the test instructions, permissible error percentage, also the
Acceptance probability. The control limit is in germany by 99 percent (about 2,58 sigma) and the
 warning limit is 95 percent(1,96 sigma).

Be aware!
In USA you have 3 sigma for control limit and 2 sigma for warning limit.
For to calculate the ppm(parts per million) it is a very big difference to calculate the
proportion of defective parts.

It uses the limits for germany!